extract tracing setup to a separate module
