+++ app-server/src/main.rs
@@ -41,17 +41,10 @@ use traces::{
 
 use cache::{Cache, in_memory::InMemoryCache, redis::RedisCache};
 use evaluators::{EVALUATORS_EXCHANGE, EVALUATORS_QUEUE, process_evaluators};
-use opentelemetry::global;
-use opentelemetry::trace::TracerProvider;
-use opentelemetry_sdk::trace::SdkTracerProvider;
 use realtime::{SseConnectionMap, cleanup_closed_connections};
-use sentry::integrations::opentelemetry as sentry_opentelemetry;
 use sodiumoxide;
 use std::{
     borrow::Cow, env, io::{self, Error}, sync::Arc, thread::{self, JoinHandle}
-};
-use tracing_subscriber::{
-    layer::SubscriberExt, util::SubscriberInitExt, EnvFilter,
 };
 
 mod agent_manager;
@@ -65,6 +58,7 @@ mod db;
 mod evaluations;
 mod evaluators;
 mod features;
+mod instrumentation;
 mod language_model;
 mod mq;
 mod names;
@@ -121,33 +115,9 @@ fn main() -> anyhow::Result<()> {
         drop(_sentry_guard);
     }
 
-    // == OpenTelemetry Tracer Provider ==
-    let tracer_provider = SdkTracerProvider::builder()
-        // Register the Sentry span processor to send OpenTelemetry spans to Sentry
-        .with_span_processor(sentry_opentelemetry::SentrySpanProcessor::new())
-        .build();
-
-    // == Tracing Subscriber with OpenTelemetry ==
-    // Create the tracer from the provider BEFORE setting it globally
-    // This ensures we have a concrete SdkTracer type, not a BoxedTracer
-    let tracer = tracer_provider.tracer("app-server");
-    
-    // Now set the global tracer provider
-    global::set_tracer_provider(tracer_provider);
-    
-    // Create environment filter (respects RUST_LOG env var)
-    let env_filter = if env::var("RUST_LOG").is_ok_and(|s| !s.is_empty()) {
-        EnvFilter::from_default_env()
-    } else {
-        EnvFilter::new("info")
-    };
-
-    // Set up the tracing subscriber with both OpenTelemetry and console output
-    tracing_subscriber::registry()
-        .with(env_filter)
-        .with(tracing_opentelemetry::layer().with_tracer(tracer))
-        .with(tracing_subscriber::fmt::layer())
-        .init();
+    if is_feature_enabled(Feature::Tracing) {
+        instrumentation::setup_tracing();
+    }
 
     let http_payload_limit: usize = env::var("HTTP_PAYLOAD_LIMIT")
         .unwrap_or(String::from("5242880")) // default to 5MB
